+++ examples_edurov/minimal/__init__.py
@@ -1,1 +1,0 @@
-# To make the folder into a package